refactor: move entires to a separate folder
