+++ src/entries/Home.tsx
@@ -1,8 +1,8 @@
 import { Divider } from 'antd'
-import * as React from 'react'
+import React from 'react'
 import { Link } from 'react-router-dom'
 
-export const Home = (): React.ReactElement<any> => {
+export const Home = (): React.ReactElement<{}> => {
 	const year = new Date().getFullYear()
 	return (
 		<div id="home">
